+++ packages/components/cascader-panel/src/index.vue
@@ -57,12 +57,12 @@ import { checkNode, getMenuIndex, sortByOriginalOrder } from './utils'
 import { CASCADER_PANEL_INJECTION_KEY } from './types'
 
 import type {
-  default as CascaderNode,
+  CascaderNode,
   CascaderNodeValue,
   CascaderOption,
   CascaderValue,
-} from './node'
-import type { ElCascaderPanelContext } from './types'
+  ElCascaderPanelContext,
+} from './types'
 import type { CascaderMenuInstance } from './instance'
 
 defineOptions({
