+++ packages/components/cascader-panel/src/node.vue
@@ -69,7 +69,7 @@ import { ArrowRight, Check, Loading } from '@element-plus/icons-vue'
 import NodeContent from './node-content'
 import { CASCADER_PANEL_INJECTION_KEY } from './types'
 
-import type { default as CascaderNode } from './node'
+import type { CascaderNode } from './types'
 import type { PropType } from 'vue'
 import type { CheckboxValueType } from '@element-plus/components/checkbox'
