+++ packages/components/cascader-panel/src/types.ts
@@ -1,13 +1,8 @@
 import type { InjectionKey, VNode } from 'vue'
-import type {
-  default as CascaderNode,
-  CascaderOption,
-  CascaderProps,
-  ExpandTrigger,
-} from './node'
+import type { Nullable } from '@element-plus/utils'
+import type { default as CascaderNode } from './node'
 
-export type { CascaderNode, CascaderOption, CascaderProps, ExpandTrigger }
-
+export type { CascaderNode, Nullable }
 export type CascaderNodeValue = string | number | Record<string, any>
 export type CascaderNodePathValue = CascaderNodeValue[]
 export type CascaderValue =
@@ -15,6 +10,7 @@ export type CascaderValue =
   | CascaderNodePathValue
   | (CascaderNodeValue | CascaderNodePathValue)[]
 export type CascaderConfig = Required<CascaderProps>
+export type ExpandTrigger = 'click' | 'hover'
 export type isDisabled = (data: CascaderOption, node: CascaderNode) => boolean
 export type isLeaf = (data: CascaderOption, node: CascaderNode) => boolean
 export type Resolve = (dataList?: CascaderOption[]) => void
@@ -24,6 +20,31 @@ export interface RenderLabelProps {
   data: CascaderOption
 }
 export type RenderLabel = (props: RenderLabelProps) => VNode | VNode[]
+export interface CascaderOption extends Record<string, unknown> {
+  label?: string
+  value?: CascaderNodeValue
+  children?: CascaderOption[]
+  disabled?: boolean
+  leaf?: boolean
+}
+
+export interface CascaderProps {
+  expandTrigger?: ExpandTrigger
+  multiple?: boolean
+  checkStrictly?: boolean
+  emitPath?: boolean
+  lazy?: boolean
+  lazyLoad?: LazyLoad
+  value?: string
+  label?: string
+  children?: string
+  disabled?: string | isDisabled
+  leaf?: string | isLeaf
+  hoverThreshold?: number
+  checkOnClickNode?: boolean
+  checkOnClickLeaf?: boolean
+  showPrefix?: boolean
+}
 
 export interface Tag {
   node?: CascaderNode
